requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.3.0
shodan==1.29.0
